@@ -28,24 +28,24 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   calendarro:
     path: ../calendarro
-  cloud_firestore: ^3.1.15
+  cloud_firestore: ^4.4.5
   cupertino_icons: ^1.0.3 
   draggable_scrollbar: ^0.1.0
-  firebase_auth: ^3.3.14
-  firebase_core: ^1.14.1
-  firebase_storage: ^10.2.12
+  firebase_auth: ^4.2.2
+  firebase_core: ^2.8.0
+  firebase_storage: ^11.0.16
   flutter_image_compress: ^1.1.0
-  flutter_launcher_icons: ^0.9.2
-  flutter_local_notifications: ^9.5.3+1
-  geolocator: ^8.2.1
-  google_fonts: ^3.0.1
+  flutter_launcher_icons: ^0.12.0
+  flutter_local_notifications: ^14.0.0-dev.2
+  geolocator: ^9.0.2
+  google_fonts: ^4.0.3
   google_maps_flutter: ^2.0.8
   image_picker: ^0.8.4
-  intl: ^0.17.0
+  intl: ^0.18.0
   location: ^4.3.0
   permission_handler: ^10.0.0
   provider: ^6.0.0
-  purchases_flutter: ^3.4.4
+  purchases_flutter: ^4.11.1
   rate_my_app: ^1.1.1
   responsive_framework: ^0.2.0
   rxdart: ^0.27.2
@@ -81,8 +81,8 @@ flutter:
   #  - images/a_dot_ham.jpeg
 
   assets:
-    - LauncherIcon.png
-    - OriginalonTransparent.png
+    - assets/LauncherIcon.png
+    - assets/OriginalonTransparent.png
     - assets/chequered-flag.png
     - assets/chequered-flagiOS.png
     - assets/bigHome.png
