@@ -11,7 +11,7 @@ class FormSubmitButton extends CustomRaisedButton {
           context: context,
           child: Text(
             text,
-            style: Theme.of(context).primaryTextTheme.headline5,
+            style: Theme.of(context).primaryTextTheme.headlineSmall,
           ),
           height: 44.0,
           textColor: Theme.of(context).colorScheme.secondary,
