@@ -22,25 +22,27 @@ class AccessOptions extends StatefulWidget {
 class _AccessOptionsState extends State<AccessOptions> {
   late Session session;
   late Database database;
-  Authorizations authorizations =
-  Authorizations(authorizedRoles: {}, authorizedNames: {}, authorizedDates: {});
+  Authorizations authorizations = Authorizations(
+      authorizedRoles: {}, authorizedNames: {}, authorizedDates: {});
   Restaurant? get restaurant => session.currentRestaurant;
   UserMessage? get message => widget.message;
   double buttonSize = 180.0;
-  bool get hasPermAuth => authorizations.authorizedRoles!.containsKey(message!.fromUid) &&
-                          authorizations.authorizedRoles![message!.fromUid] == ROLE_STAFF;
-  bool get hasVenueAuth => authorizations.authorizedRoles!.containsKey(message!.fromUid) &&
-                           authorizations.authorizedRoles![message!.fromUid] == ROLE_VENUE;
+  bool get hasPermAuth =>
+      authorizations.authorizedRoles!.containsKey(message!.fromUid) &&
+      authorizations.authorizedRoles![message!.fromUid] == ROLE_STAFF;
+  bool get hasVenueAuth =>
+      authorizations.authorizedRoles!.containsKey(message!.fromUid) &&
+      authorizations.authorizedRoles![message!.fromUid] == ROLE_VENUE;
 
   void _changeVenueAuthorization() async {
-    List<dynamic> authorizedIntDates = authorizations.authorizedDates![message!.fromUid] ?? [];
+    List<dynamic> authorizedIntDates =
+        authorizations.authorizedDates![message!.fromUid] ?? [];
     final authorizedDates = await Navigator.of(context).push(
       MaterialPageRoute<List<DateTime>>(
         fullscreenDialog: false,
-        builder: (context) =>
-            AuthorizedDates(
-              authorizedIntDates: authorizedIntDates,
-            ),
+        builder: (context) => AuthorizedDates(
+          authorizedIntDates: authorizedIntDates,
+        ),
       ),
     );
     if (authorizedDates != null && authorizedDates.length > 0) {
@@ -58,7 +60,8 @@ class _AccessOptionsState extends State<AccessOptions> {
       authorizations.authorizedNames!
           .putIfAbsent(message!.fromUid, () => message!.fromName);
       if (authorizations.authorizedDates!.containsKey(message!.fromUid)) {
-        authorizations.authorizedDates!.update(message!.fromUid, (value) => authorizedIntDates);
+        authorizations.authorizedDates!
+            .update(message!.fromUid, (value) => authorizedIntDates);
       } else {
         authorizations.authorizedDates!
             .putIfAbsent(message!.fromUid, () => authorizedIntDates);
@@ -66,7 +69,7 @@ class _AccessOptionsState extends State<AccessOptions> {
       setState(() {
         message!.authFlag = true;
       });
-   } else {
+    } else {
       authorizations.authorizedRoles!.remove(message!.fromUid);
       authorizations.authorizedNames!.remove(message!.fromUid);
       authorizations.authorizedDates!.remove(message!.fromUid);
@@ -113,32 +116,24 @@ class _AccessOptionsState extends State<AccessOptions> {
         type: message!.type,
         authFlag: message!.authFlag,
         delivered: true,
-        attendedFlag: true
-    );
+        attendedFlag: true);
     database.setMessageDetails(readMessage);
   }
 
   List<Widget> _buildContents(BuildContext context) {
     return [
-      Text(
-          message!.type!,
-          style: Theme.of(context).textTheme.headline5
-      ),
+      Text(message!.type!, style: Theme.of(context).textTheme.headlineSmall),
       SizedBox(
         height: 16.0,
       ),
-      Text(
-          'Requested by ${message!.fromName}',
+      Text('Requested by ${message!.fromName}',
           textAlign: TextAlign.center,
-          style: Theme.of(context).textTheme.headline6
-      ),
+          style: Theme.of(context).textTheme.titleLarge),
       SizedBox(
         height: 16.0,
       ),
-      Text(
-          'Select access type',
-          style: Theme.of(context).textTheme.headline5
-      ),
+      Text('Select access type',
+          style: Theme.of(context).textTheme.headlineSmall),
       SizedBox(
         height: 32.0,
       ),
@@ -156,18 +151,19 @@ class _AccessOptionsState extends State<AccessOptions> {
               Icons.account_box,
               size: 36.0,
             ),
-            SizedBox(height: 8.0,),
+            SizedBox(
+              height: 8.0,
+            ),
             Text(
               'Permanent Staff Access',
-              style: Theme.of(context).textTheme.headline6,
+              style: Theme.of(context).textTheme.titleLarge,
               textAlign: TextAlign.center,
             ),
             CupertinoSwitch(
-              value: hasPermAuth,
-              onChanged: (flag) {
-                _changePermAuthorization(flag);
-              }
-            ),
+                value: hasPermAuth,
+                onChanged: (flag) {
+                  _changePermAuthorization(flag);
+                }),
           ],
         ),
       ),
@@ -178,8 +174,8 @@ class _AccessOptionsState extends State<AccessOptions> {
         height: buttonSize,
         width: buttonSize,
         decoration: BoxDecoration(
-            border: Border.all(),
-            borderRadius: BorderRadius.circular(12),
+          border: Border.all(),
+          borderRadius: BorderRadius.circular(12),
         ),
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
@@ -188,18 +184,19 @@ class _AccessOptionsState extends State<AccessOptions> {
               Icons.calendar_today,
               size: 36.0,
             ),
-            SizedBox(height: 8.0,),
+            SizedBox(
+              height: 8.0,
+            ),
             Text(
               'Sales Access\n(Selected Dates)',
-              style: Theme.of(context).textTheme.headline6,
+              style: Theme.of(context).textTheme.titleLarge,
               textAlign: TextAlign.center,
             ),
             CupertinoSwitch(
                 value: hasVenueAuth,
                 onChanged: (flag) {
                   _changeVenueAuthorization();
-                }
-            ),
+                }),
           ],
         ),
       ),
@@ -214,38 +211,39 @@ class _AccessOptionsState extends State<AccessOptions> {
         appBar: AppBar(
           title: Text(
             'Access Options',
-            style: TextStyle(color: Theme.of(context).appBarTheme.backgroundColor),
+            style:
+                TextStyle(color: Theme.of(context).appBarTheme.backgroundColor),
           ),
         ),
         body: FutureBuilder<List<Authorizations>>(
-          future: database.authorizationsSnapshot(),
-          builder: (context, snapshot) {
-            if (snapshot.connectionState != ConnectionState.waiting &&
-                snapshot.hasData) {
-              final authorizationsList = snapshot.data!;
-              if (authorizationsList.length > 0) {
-                authorizationsList.forEach((authorization) {
-                  if (authorization.id == message!.restaurantId) {
-                    authorizations = authorization;
-                  }
-                });
-              }
-              return SingleChildScrollView(
-                child: Center(
-                  child: Padding(
-                    padding: const EdgeInsets.all(32.0),
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: _buildContents(context),
+            future: database.authorizationsSnapshot(),
+            builder: (context, snapshot) {
+              if (snapshot.connectionState != ConnectionState.waiting &&
+                  snapshot.hasData) {
+                final authorizationsList = snapshot.data!;
+                if (authorizationsList.length > 0) {
+                  authorizationsList.forEach((authorization) {
+                    if (authorization.id == message!.restaurantId) {
+                      authorizations = authorization;
+                    }
+                  });
+                }
+                return SingleChildScrollView(
+                  child: Center(
+                    child: Padding(
+                      padding: const EdgeInsets.all(32.0),
+                      child: Column(
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        children: _buildContents(context),
+                      ),
                     ),
                   ),
-                ),
-              );
-            } else {
-              return Center(child: PlatformProgressIndicator(),);
-            }
-          }
-        )
-    );
+                );
+              } else {
+                return Center(
+                  child: PlatformProgressIndicator(),
+                );
+              }
+            }));
   }
 }
