@@ -8,7 +8,7 @@ class VenueAuthorizationPage extends StatelessWidget {
         padding: const EdgeInsets.all(8.0),
         child: Text(
           'No authorized dates found. Please contact the restaurant manager.',
-          style: Theme.of(context).textTheme.headline5,
+          style: Theme.of(context).textTheme.headlineSmall,
           textAlign: TextAlign.center,
         ),
       ),
