@@ -13,7 +13,8 @@ class SignInPage extends StatelessWidget {
   final bool? allowAnonymousSignIn;
   final bool? convertAnonymous;
 
-  const SignInPage({Key? key, this.allowAnonymousSignIn, this.convertAnonymous}) : super(key: key);
+  const SignInPage({Key? key, this.allowAnonymousSignIn, this.convertAnonymous})
+      : super(key: key);
 
   void _showSignInError(BuildContext context, PlatformException exception) {
     PlatformExceptionAlertDialog(
@@ -35,7 +36,9 @@ class SignInPage extends StatelessWidget {
     Navigator.of(context).push(
       MaterialPageRoute<void>(
         fullscreenDialog: true,
-        builder: (context) => EmailSignInPage(convertAnonymous: convertAnonymous,),
+        builder: (context) => EmailSignInPage(
+          convertAnonymous: convertAnonymous,
+        ),
       ),
     );
   }
@@ -83,18 +86,18 @@ class SignInPage extends StatelessWidget {
           SizedBox(height: 24.0),
           // ANON
           if (allowAnonymousSignIn!)
-          TextButton(
-            child: Text(
+            TextButton(
+              child: Text(
                 'I\'ll sign-in later',
-              style: Theme.of(context).textTheme.headline5,
+                style: Theme.of(context).textTheme.headlineSmall,
+              ),
+              onPressed: () => _signInAnonymously(context),
             ),
-            onPressed: () => _signInAnonymously(context),
-          ),
           if (convertAnonymous!)
             TextButton(
               child: Text(
                 'You\'re currently an anonymous user.\nYou need to sign-in or create an account. \nTap here to learn why.',
-                style: Theme.of(context).textTheme.bodyText2,
+                style: Theme.of(context).textTheme.bodyMedium,
                 textAlign: TextAlign.center,
               ),
               onPressed: () => _displayWhy(context),
@@ -106,13 +109,17 @@ class SignInPage extends StatelessWidget {
   }
 
   void _displayWhy(BuildContext context) {
-    String reason = 'If you want to continue you need to register an account for the following reason(s): \n\n- We want to keep a clean database and discourage spammers';
+    String reason =
+        'If you want to continue you need to register an account for the following reason(s): \n\n- We want to keep a clean database and discourage spammers';
     if (FlavourConfig.isManager()) {
-      reason = reason + '\n\n- You will be able to purchase order bundles to unlock your restaurant orders\n\n- You will be able to add images to your restaurant gallery\n\n- You will be able to grant access to restaurant orders to NM Staff users';
+      reason = reason +
+          '\n\n- You will be able to purchase order bundles to unlock your restaurant orders\n\n- You will be able to add images to your restaurant gallery\n\n- You will be able to grant access to restaurant orders to NM Staff users';
     } else if (FlavourConfig.isStaff()) {
-      reason = reason + '\n- Restaurant managers need to know who you are to grant you restaurant access';
+      reason = reason +
+          '\n- Restaurant managers need to know who you are to grant you restaurant access';
     } else {
-      reason = reason + '\n- Your contact details are needed for order deliveries';
+      reason =
+          reason + '\n- Your contact details are needed for order deliveries';
     }
     PlatformAlertDialog(
       title: 'Why you need an account',
@@ -125,8 +132,7 @@ class SignInPage extends StatelessWidget {
     return Text(
       'Welcome',
       textAlign: TextAlign.center,
-      style: Theme.of(context).primaryTextTheme.headline4,
+      style: Theme.of(context).primaryTextTheme.headlineMedium,
     );
   }
-
 }
