@@ -11,16 +11,11 @@ class GreenTheme {
     //accentColorBrightness: Brightness.light,
     canvasColor: Color(0xfffafafa),
     scaffoldBackgroundColor: Color(0xfffafafa),
-    bottomAppBarColor: Color(0xffffffff),
+    bottomAppBarTheme: BottomAppBarTheme(color: Color(0xffffffff)),
     cardColor: Color(0xffffffff),
     dividerColor: Color(0x1f000000),
     highlightColor: Color(0x66bcbcbc),
     splashColor: Color(0x66c8c8c8),
-    selectedRowColor: Color(0xfff5f5f5),
-    unselectedWidgetColor: Color(0x8a000000),
-    disabledColor: Color(0x61000000),
-    //buttonColor: Color(0xffe0e0e0),
-    toggleableActiveColor: Color(0xff7cb342),
     secondaryHeaderColor: Color(0xfff1f8e9),
     //textSelectionColor: Color(0xffc5e1a5),
     //cursorColor: Color(0xff4285f4),
@@ -34,14 +29,15 @@ class GreenTheme {
       textTheme: ButtonTextTheme.normal,
       minWidth: 88.0,
       height: 36.0,
-      padding: EdgeInsets.only(
-          top: 0.0, bottom: 0.0, left: 16.0, right: 16.0),
+      padding: EdgeInsets.only(top: 0.0, bottom: 0.0, left: 16.0, right: 16.0),
       shape: RoundedRectangleBorder(
         side: BorderSide(
-          color: Color(0xff000000), width: 0.0, style: BorderStyle.none,),
+          color: Color(0xff000000),
+          width: 0.0,
+          style: BorderStyle.none,
+        ),
         borderRadius: BorderRadius.all(Radius.circular(2.0)),
-      )
-      ,
+      ),
       alignedDropdown: false,
       buttonColor: Color(0xffe0e0e0),
       disabledColor: Color(0x61000000),
@@ -62,79 +58,79 @@ class GreenTheme {
       ),
     ),
     textTheme: TextTheme(
-      headline1: TextStyle(
+      displayLarge: TextStyle(
         color: Color(0x8a000000),
         fontSize: 96.0,
         fontWeight: FontWeight.w300,
         fontStyle: FontStyle.normal,
       ),
-      headline2: TextStyle(
+      displayMedium: TextStyle(
         color: Color(0x8a000000),
         fontSize: 60.0,
         fontWeight: FontWeight.w300,
         fontStyle: FontStyle.normal,
       ),
-      headline3: TextStyle(
+      displaySmall: TextStyle(
         color: Color(0x8a000000),
         fontSize: 48.0,
         fontWeight: FontWeight.w400,
         fontStyle: FontStyle.normal,
       ),
-      headline4: TextStyle(
+      headlineMedium: TextStyle(
         color: Color(0x8a000000),
         fontSize: 34.0,
         fontWeight: FontWeight.w400,
         fontStyle: FontStyle.normal,
       ),
-      headline5: TextStyle(
+      headlineSmall: TextStyle(
         color: Color(0xdd000000),
         fontSize: 24.0,
         fontWeight: FontWeight.w400,
         fontStyle: FontStyle.normal,
       ),
-      headline6: TextStyle(
+      titleLarge: TextStyle(
         color: Color(0xdd000000),
         fontSize: 20.0,
         fontWeight: FontWeight.w500,
         fontStyle: FontStyle.normal,
       ),
-      subtitle1: TextStyle(
+      titleMedium: TextStyle(
         color: Color(0xdd000000),
         fontSize: 16.0,
         fontWeight: FontWeight.w400,
         fontStyle: FontStyle.normal,
       ),
-      bodyText1: TextStyle(
+      bodyLarge: TextStyle(
         color: Color(0xdd000000),
         fontSize: 14.0,
         fontWeight: FontWeight.w400,
         fontStyle: FontStyle.normal,
       ),
-      bodyText2: TextStyle(
+      bodyMedium: TextStyle(
         color: Color(0xdd000000),
         fontSize: 16.0,
         fontWeight: FontWeight.w400,
         fontStyle: FontStyle.normal,
       ),
-      caption: TextStyle(
+      bodySmall: TextStyle(
         color: Color(0x8a000000),
         fontSize: 12.0,
         fontWeight: FontWeight.w400,
         fontStyle: FontStyle.normal,
       ),
-      button: TextStyle(
+      labelLarge: TextStyle(
         color: Color(0xdd000000),
         fontSize: 14.0,
         fontWeight: FontWeight.w500,
         fontStyle: FontStyle.normal,
       ),
-      subtitle2: TextStyle(
+      titleSmall: TextStyle(
         color: Color(0xff000000),
         fontSize: 14.0,
         fontWeight: FontWeight.w500,
         fontStyle: FontStyle.normal,
       ),
-      overline: TextStyle(
+      labelSmall: TextStyle(
         color: Color(0xff000000),
         fontSize: 10.0,
         fontWeight: FontWeight.w400,
@@ -142,165 +138,85 @@ class GreenTheme {
       ),
     ),
     primaryTextTheme: TextTheme(
-      headline1: TextStyle(
+      displayLarge: TextStyle(
         color: Color(0x8a000000),
         fontSize: 96.0,
         fontWeight: FontWeight.w300,
         fontStyle: FontStyle.normal,
       ),
-      headline2: TextStyle(
+      displayMedium: TextStyle(
         color: Color(0x8a000000),
         fontSize: 60.0,
         fontWeight: FontWeight.w300,
         fontStyle: FontStyle.normal,
       ),
-      headline3: TextStyle(
+      displaySmall: TextStyle(
         color: Color(0x8a000000),
         fontSize: 48.0,
         fontWeight: FontWeight.w400,
         fontStyle: FontStyle.normal,
       ),
-      headline4: TextStyle(
+      headlineMedium: TextStyle(
         color: Color(0x8a000000),
         fontSize: 34.0,
         fontWeight: FontWeight.w400,
         fontStyle: FontStyle.normal,
       ),
-      headline5: TextStyle(
+      headlineSmall: TextStyle(
         color: Color(0xdd000000),
         fontSize: 24.0,
         fontWeight: FontWeight.w400,
         fontStyle: FontStyle.normal,
       ),
-      headline6: TextStyle(
+      titleLarge: TextStyle(
         color: Color(0xdd000000),
         fontSize: 20.0,
         fontWeight: FontWeight.w500,
         fontStyle: FontStyle.normal,
       ),
-      subtitle1: TextStyle(
+      titleMedium: TextStyle(
         color: Color(0xdd000000),
         fontSize: 16.0,
         fontWeight: FontWeight.w400,
         fontStyle: FontStyle.normal,
       ),
-      bodyText1: TextStyle(
+      bodyLarge: TextStyle(
         color: Color(0xdd000000),
         fontSize: 14.0,
         fontWeight: FontWeight.w400,
         fontStyle: FontStyle.normal,
       ),
-      bodyText2: TextStyle(
+      bodyMedium: TextStyle(
         color: Color(0xdd000000),
         fontSize: 16.0,
         fontWeight: FontWeight.w400,
         fontStyle: FontStyle.normal,
       ),
-      caption: TextStyle(
+      bodySmall: TextStyle(
         color: Color(0x8a000000),
         fontSize: 12.0,
         fontWeight: FontWeight.w400,
         fontStyle: FontStyle.normal,
       ),
-      button: TextStyle(
+      labelLarge: TextStyle(
         color: Color(0xdd000000),
         fontSize: 14.0,
         fontWeight: FontWeight.w500,
         fontStyle: FontStyle.normal,
       ),
-      subtitle2: TextStyle(
+      titleSmall: TextStyle(
         color: Color(0xff000000),
         fontSize: 14.0,
         fontWeight: FontWeight.w500,
         fontStyle: FontStyle.normal,
       ),
-      overline: TextStyle(
+      labelSmall: TextStyle(
         color: Color(0xff000000),
         fontSize: 10.0,
         fontWeight: FontWeight.w400,
         fontStyle: FontStyle.normal,
       ),
     ),
-    // accentTextTheme: TextTheme(
-    //   headline1: TextStyle(
-    //     color: Color(0x8a000000),
-    //     fontSize: 96.0,
-    //     fontWeight: FontWeight.w300,
-    //     fontStyle: FontStyle.normal,
-    //   ),
-    //   headline2: TextStyle(
-    //     color: Color(0x8a000000),
-    //     fontSize: 60.0,
-    //     fontWeight: FontWeight.w300,
-    //     fontStyle: FontStyle.normal,
-    //   ),
-    //   headline3: TextStyle(
-    //     color: Color(0x8a000000),
-    //     fontSize: 48.0,
-    //     fontWeight: FontWeight.w400,
-    //     fontStyle: FontStyle.normal,
-    //   ),
-    //   headline4: TextStyle(
-    //     color: Color(0x8a000000),
-    //     fontSize: 34.0,
-    //     fontWeight: FontWeight.w400,
-    //     fontStyle: FontStyle.normal,
-    //   ),
-    //   headline5: TextStyle(
-    //     color: Color(0xdd000000),
-    //     fontSize: 24.0,
-    //     fontWeight: FontWeight.w400,
-    //     fontStyle: FontStyle.normal,
-    //   ),
-    //   headline6: TextStyle(
-    //     color: Color(0xdd000000),
-    //     fontSize: 20.0,
-    //     fontWeight: FontWeight.w500,
-    //     fontStyle: FontStyle.normal,
-    //   ),
-    //   subtitle1: TextStyle(
-    //     color: Color(0xdd000000),
-    //     fontSize: 16.0,
-    //     fontWeight: FontWeight.w400,
-    //     fontStyle: FontStyle.normal,
-    //   ),
-    //   bodyText1: TextStyle(
-    //     color: Color(0xdd000000),
-    //     fontSize: 14.0,
-    //     fontWeight: FontWeight.w400,
-    //     fontStyle: FontStyle.normal,
-    //   ),
-    //   bodyText2: TextStyle(
-    //     color: Color(0xdd000000),
-    //     fontSize: 16.0,
-    //     fontWeight: FontWeight.w400,
-    //     fontStyle: FontStyle.normal,
-    //   ),
-    //   caption: TextStyle(
-    //     color: Color(0x8a000000),
-    //     fontSize: 12.0,
-    //     fontWeight: FontWeight.w400,
-    //     fontStyle: FontStyle.normal,
-    //   ),
-    //   button: TextStyle(
-    //     color: Color(0xdd000000),
-    //     fontSize: 14.0,
-    //     fontWeight: FontWeight.w500,
-    //     fontStyle: FontStyle.normal,
-    //   ),
-    //   subtitle2: TextStyle(
-    //     color: Color(0xff000000),
-    //     fontSize: 14.0,
-    //     fontWeight: FontWeight.w500,
-    //     fontStyle: FontStyle.normal,
-    //   ),
-    //   overline: TextStyle(
-    //     color: Color(0xff000000),
-    //     fontSize: 10.0,
-    //     fontWeight: FontWeight.w400,
-    //     fontStyle: FontStyle.normal,
-    //   ),
-    // ),
     inputDecorationTheme: InputDecorationTheme(
       labelStyle: TextStyle(
         color: Color(0xdd000000),
@@ -329,8 +245,8 @@ class GreenTheme {
       errorMaxLines: null,
       floatingLabelBehavior: FloatingLabelBehavior.never,
       isDense: false,
-      contentPadding: EdgeInsets.only(
-          top: 12.0, bottom: 12.0, left: 0.0, right: 0.0),
+      contentPadding:
+          EdgeInsets.only(top: 12.0, bottom: 12.0, left: 0.0, right: 0.0),
       isCollapsed: false,
       prefixStyle: TextStyle(
         color: Color(0xdd000000),
@@ -354,32 +270,50 @@ class GreenTheme {
       fillColor: Color(0x00000000),
       errorBorder: UnderlineInputBorder(
         borderSide: BorderSide(
-          color: Color(0xff000000), width: 1.0, style: BorderStyle.solid,),
+          color: Color(0xff000000),
+          width: 1.0,
+          style: BorderStyle.solid,
+        ),
         borderRadius: BorderRadius.all(Radius.circular(4.0)),
       ),
       focusedBorder: UnderlineInputBorder(
         borderSide: BorderSide(
-          color: Color(0xff000000), width: 1.0, style: BorderStyle.solid,),
+          color: Color(0xff000000),
+          width: 1.0,
+          style: BorderStyle.solid,
+        ),
         borderRadius: BorderRadius.all(Radius.circular(4.0)),
       ),
       focusedErrorBorder: UnderlineInputBorder(
         borderSide: BorderSide(
-          color: Color(0xff000000), width: 1.0, style: BorderStyle.solid,),
+          color: Color(0xff000000),
+          width: 1.0,
+          style: BorderStyle.solid,
+        ),
         borderRadius: BorderRadius.all(Radius.circular(4.0)),
       ),
       disabledBorder: UnderlineInputBorder(
         borderSide: BorderSide(
-          color: Color(0xff000000), width: 1.0, style: BorderStyle.solid,),
+          color: Color(0xff000000),
+          width: 1.0,
+          style: BorderStyle.solid,
+        ),
         borderRadius: BorderRadius.all(Radius.circular(4.0)),
       ),
       enabledBorder: UnderlineInputBorder(
         borderSide: BorderSide(
-          color: Color(0xff000000), width: 1.0, style: BorderStyle.solid,),
+          color: Color(0xff000000),
+          width: 1.0,
+          style: BorderStyle.solid,
+        ),
         borderRadius: BorderRadius.all(Radius.circular(4.0)),
       ),
       border: UnderlineInputBorder(
         borderSide: BorderSide(
-          color: Color(0xff000000), width: 1.0, style: BorderStyle.solid,),
+          color: Color(0xff000000),
+          width: 1.0,
+          style: BorderStyle.solid,
+        ),
         borderRadius: BorderRadius.all(Radius.circular(4.0)),
       ),
     ),
@@ -434,8 +368,8 @@ class GreenTheme {
       brightness: Brightness.light,
       deleteIconColor: Color(0xde000000),
       disabledColor: Color(0x0c000000),
-      labelPadding: EdgeInsets.only(
-          top: 0.0, bottom: 0.0, left: 8.0, right: 8.0),
+      labelPadding:
+          EdgeInsets.only(top: 0.0, bottom: 0.0, left: 8.0, right: 8.0),
       labelStyle: TextStyle(
         color: Color(0xde000000),
         fontSize: 14.0,
@@ -451,17 +385,21 @@ class GreenTheme {
       ),
       secondarySelectedColor: Color(0x3d8bc34a),
       selectedColor: Color(0x3d000000),
-      shape: StadiumBorder(side: BorderSide(
-        color: Color(0xff000000), width: 0.0, style: BorderStyle.none,)),
+      shape: StadiumBorder(
+          side: BorderSide(
+        color: Color(0xff000000),
+        width: 0.0,
+        style: BorderStyle.none,
+      )),
     ),
     dialogTheme: DialogTheme(
         shape: RoundedRectangleBorder(
-          side: BorderSide(color: Color(0xff000000),
-            width: 0.0,
-            style: BorderStyle.none,),
-          borderRadius: BorderRadius.all(Radius.circular(0.0)),
-        )
-
-    ),
+      side: BorderSide(
+        color: Color(0xff000000),
+        width: 0.0,
+        style: BorderStyle.none,
+      ),
+      borderRadius: BorderRadius.all(Radius.circular(0.0)),
+    )),
   );
 }
