@@ -6,8 +6,8 @@ class CustomRaisedButton extends StatelessWidget {
     this.child,
     this.textColor,
     this.color,
-    this.borderRadius : 6.0,
-    this.height : 50.0,
+    this.borderRadius = 6.0,
+    this.height = 50.0,
     this.width,
     this.onPressed,
   });
@@ -27,28 +27,30 @@ class CustomRaisedButton extends StatelessWidget {
       height: height,
       width: width,
       child: ElevatedButton(
-          onPressed: onPressed,
-          child: child,
-          style: ButtonStyle(
-            backgroundColor: MaterialStateProperty.resolveWith<Color?>(
-              (Set<MaterialState> states) {
-                if (states.contains(MaterialState.pressed))
-                  return Theme.of(context).buttonTheme.colorScheme!.background; //color;
-                else if (states.contains(MaterialState.disabled))
-                  return Theme.of(context).buttonTheme.colorScheme!.background; //color;
-                return null; // Use the component's default.
-              },
-            ),
-            shape: MaterialStateProperty.all<RoundedRectangleBorder>(
-              RoundedRectangleBorder(
-                borderRadius: BorderRadius.all(
-                  Radius.circular(
-                    borderRadius,
-                  )
-                )
-              )
-            ),
+        onPressed: onPressed,
+        child: child,
+        style: ButtonStyle(
+          backgroundColor: MaterialStateProperty.resolveWith<Color?>(
+            (Set<MaterialState> states) {
+              if (states.contains(MaterialState.pressed))
+                return Theme.of(context)
+                    .buttonTheme
+                    .colorScheme!
+                    .background; //color;
+              else if (states.contains(MaterialState.disabled))
+                return Theme.of(context)
+                    .buttonTheme
+                    .colorScheme!
+                    .background; //color;
+              return null; // Use the component's default.
+            },
           ),
+          shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+              RoundedRectangleBorder(
+                  borderRadius: BorderRadius.all(Radius.circular(
+            borderRadius,
+          )))),
+        ),
       ),
     );
   }
