@@ -24,36 +24,32 @@ class CupertinoHomeScaffold extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final session = Provider.of<Session>(context);
-    return Stack(
-      children: <Widget>[
-          CupertinoTabScaffold(
-            tabBar: CupertinoTabBar(
-              backgroundColor: FlavourConfig.isManager()
-                  ?  Theme.of(context).primaryColor
-                  : Theme.of(context).backgroundColor,
-              activeColor: Theme.of(context).colorScheme.secondary,
-              items: RoleEnumBase.itemsForRole(context, currentTab, roleTabItems),
-              onTap: (index) => onSelectTab(roleTabItems.roleEnumList![index]),
-            ),
-            resizeToAvoidBottomInset: false,
-            tabBuilder: (context, index) {
-              final item = roleTabItems.roleEnumList![index];
-              return CupertinoTabView(
-                builder: (context) => widgetBuilders[item]!(context),
-                navigatorKey: navigatorKeys[item],
-            );
-          },
+    return Stack(children: <Widget>[
+      CupertinoTabScaffold(
+        tabBar: CupertinoTabBar(
+          backgroundColor: FlavourConfig.isManager()
+              ? Theme.of(context).primaryColor
+              : Theme.of(context).colorScheme.background,
+          activeColor: Theme.of(context).colorScheme.secondary,
+          items: RoleEnumBase.itemsForRole(context, currentTab, roleTabItems),
+          onTap: (index) => onSelectTab(roleTabItems.roleEnumList![index]),
         ),
-        if (FlavourConfig.isManager())
-          StreamBuilder<int>(
+        resizeToAvoidBottomInset: false,
+        tabBuilder: (context, index) {
+          final item = roleTabItems.roleEnumList![index];
+          return CupertinoTabView(
+            builder: (context) => widgetBuilders[item]!(context),
+            navigatorKey: navigatorKeys[item],
+          );
+        },
+      ),
+      if (FlavourConfig.isManager())
+        StreamBuilder<int>(
             stream: session.messageCounterObservable,
             builder: (context, snapshot) {
               if (snapshot.hasData && snapshot.data! > 0) {
                 return Positioned(
-                  right: MediaQuery
-                      .of(context)
-                      .size
-                      .width / 2 - 35,
+                  right: MediaQuery.of(context).size.width / 2 - 35,
                   bottom: 25,
                   child: Container(
                     height: 20.0,
@@ -62,8 +58,7 @@ class CupertinoHomeScaffold extends StatelessWidget {
                     margin: EdgeInsets.only(top: 35.0, right: 5.0),
                     decoration: BoxDecoration(
                         borderRadius: BorderRadius.circular(12.0),
-                        color: Colors.red
-                    ),
+                        color: Colors.red),
                     child: Text(
                       snapshot.data.toString(),
                       style: const TextStyle(
@@ -75,18 +70,15 @@ class CupertinoHomeScaffold extends StatelessWidget {
                 );
               } else {
                 return Positioned(
-                  right: MediaQuery
-                      .of(context)
-                      .size
-                      .width / 2 - 35,
+                  right: MediaQuery.of(context).size.width / 2 - 35,
                   bottom: 25,
-                  child: Container(height: 20.0, width: 20.0,),
+                  child: Container(
+                    height: 20.0,
+                    width: 20.0,
+                  ),
                 );
               }
-            }
-          ),
-      ]
-    );
+            }),
+    ]);
   }
-
 }
