@@ -17,7 +17,13 @@ class OrderHistory extends StatefulWidget {
   final bool? showLocked;
   final bool? showActive;
 
-  const OrderHistory({Key? key, this.stream, this.restaurantName, this.showLocked, this.showActive}) : super(key: key);
+  const OrderHistory(
+      {Key? key,
+      this.stream,
+      this.restaurantName,
+      this.showLocked,
+      this.showActive})
+      : super(key: key);
 
   @override
   _OrderHistoryState createState() => _OrderHistoryState();
@@ -44,93 +50,97 @@ class _OrderHistoryState extends State<OrderHistory> {
     return StreamBuilder<List<Order>>(
       stream: stream,
       builder: (context, snapshot) {
-      blockedOrders = snapshot.data;
-      return ListItemsBuilder<Order>(
-        title: 'Orders',
-        message: 'There are no $messageString orders',
-        snapshot: snapshot,
-        itemBuilder: (context, order) {
-          return Card(
-            color: _determineTileColor(order),
-            margin: EdgeInsets.all(12.0),
-            child: ListTile(
-              isThreeLine: true,
-              leading: _determineIcon(order),
-              title: Padding(
-                padding: const EdgeInsets.only(top: 8.0),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      'Order # ${order.orderNumber}',
-                        style: Theme.of(context).textTheme.headline5,
-                    ),
-                    Text(
-                      '${order.restaurantName}, ${order.orderItems!.length} items',
-                      //style: Theme.of(context).textTheme.subtitle2,
-                    ),
-                  ],
-                ),
-              ),
-              subtitle: Padding(
-                padding: const EdgeInsets.only(bottom: 4.0),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: <Widget>[
-                    Padding(
-                      padding:
-                      const EdgeInsets.only(top: 4.0, bottom: 4.0),
-                      child: Text(
-                        Format.formatDateTime(order.timestamp!.toInt()),
-                      ),
-                    ),
-                    Padding(
-                      padding:
-                      const EdgeInsets.only(bottom: 4.0),
-                      child: Text(
-                        order.statusString,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              trailing: Text(
-                f.format(order.orderTotal - (order.orderTotal * order.discount!) + order.tip!),
-                style: Theme.of(context).textTheme.headline6,
-              ),
-              //onTap: role != ROLE_PATRON && order.isBlocked ? null : () => _viewOrder(context, order),
-              onTap: () async {
-                if (!order.isBlocked! || FlavourConfig.isPatron()) {
-                  _viewOrder(context, order);
-                } else {
-                  if (FlavourConfig.isManager()) {
-                    if (!widget.showLocked!) {
-                      await PlatformExceptionAlertDialog(
-                        title: 'Locked order',
-                        exception: PlatformException(
-                          code: 'LOCKED_ORDER',
-                          message: 'Please go to your profile page and buy a bundle to unlock your orders.',
-                          details: 'Please go to your profile page and buy a bundle to unlock your orders.',
+        blockedOrders = snapshot.data;
+        return ListItemsBuilder<Order>(
+            title: 'Orders',
+            message: 'There are no $messageString orders',
+            snapshot: snapshot,
+            itemBuilder: (context, order) {
+              return Card(
+                color: _determineTileColor(order),
+                margin: EdgeInsets.all(12.0),
+                child: ListTile(
+                  isThreeLine: true,
+                  leading: _determineIcon(order),
+                  title: Padding(
+                    padding: const EdgeInsets.only(top: 8.0),
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Text(
+                          'Order # ${order.orderNumber}',
+                          style: Theme.of(context).textTheme.headlineSmall,
                         ),
-                      ).show(context);
+                        Text(
+                          '${order.restaurantName}, ${order.orderItems!.length} items',
+                          //style: Theme.of(context).textTheme.subtitle2,
+                        ),
+                      ],
+                    ),
+                  ),
+                  subtitle: Padding(
+                    padding: const EdgeInsets.only(bottom: 4.0),
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: <Widget>[
+                        Padding(
+                          padding: const EdgeInsets.only(top: 4.0, bottom: 4.0),
+                          child: Text(
+                            Format.formatDateTime(order.timestamp!.toInt()),
+                          ),
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 4.0),
+                          child: Text(
+                            order.statusString,
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+                  trailing: Text(
+                    f.format(order.orderTotal -
+                        (order.orderTotal * order.discount!) +
+                        order.tip!),
+                    style: Theme.of(context).textTheme.titleLarge,
+                  ),
+                  //onTap: role != ROLE_PATRON && order.isBlocked ? null : () => _viewOrder(context, order),
+                  onTap: () async {
+                    if (!order.isBlocked! || FlavourConfig.isPatron()) {
+                      _viewOrder(context, order);
                     } else {
-                      _unlockOrders(context);
+                      if (FlavourConfig.isManager()) {
+                        if (!widget.showLocked!) {
+                          await PlatformExceptionAlertDialog(
+                            title: 'Locked order',
+                            exception: PlatformException(
+                              code: 'LOCKED_ORDER',
+                              message:
+                                  'Please go to your profile page and buy a bundle to unlock your orders.',
+                              details:
+                                  'Please go to your profile page and buy a bundle to unlock your orders.',
+                            ),
+                          ).show(context);
+                        } else {
+                          _unlockOrders(context);
+                        }
+                      } else if (FlavourConfig.isStaff()) {
+                        await PlatformExceptionAlertDialog(
+                          title: 'Locked order',
+                          exception: PlatformException(
+                            code: 'LOCKED_ORDER',
+                            message:
+                                'Order locked. Please notify your restaurant manager.',
+                            details:
+                                'Order locked. Please notify your restaurant manager.',
+                          ),
+                        ).show(context);
+                      }
                     }
-                  } else if (FlavourConfig.isStaff()) {
-                    await PlatformExceptionAlertDialog(
-                      title: 'Locked order',
-                      exception: PlatformException(
-                        code: 'LOCKED_ORDER',
-                        message:  'Order locked. Please notify your restaurant manager.',
-                        details:  'Order locked. Please notify your restaurant manager.',
-                      ),
-                    ).show(context);
-                  }
-                }
-              },
-            ),
-          );
-        });
+                  },
+                ),
+              );
+            });
       },
     );
   }
@@ -169,7 +179,7 @@ class _OrderHistoryState extends State<OrderHistory> {
   Icon _determineIcon(Order order) {
     Icon tileIcon;
     if (!FlavourConfig.isPatron() && order.isBlocked!) {
-      tileIcon =  Icon(Icons.lock);
+      tileIcon = Icon(Icons.lock);
     } else {
       switch (order.status) {
         case ORDER_PLACED:
@@ -181,7 +191,7 @@ class _OrderHistoryState extends State<OrderHistory> {
           break;
         case ORDER_READY:
           tileIcon = Icon(Icons.check);
-        break;
+          break;
         case ORDER_CANCELLED:
           tileIcon = Icon(Icons.delete_forever);
           break;
@@ -198,10 +208,9 @@ class _OrderHistoryState extends State<OrderHistory> {
       MaterialPageRoute<void>(
           fullscreenDialog: false,
           builder: (context) => ViewOrder.create(
-            context: context,
-            order: order,
-          )
-      ),
+                context: context,
+                order: order,
+              )),
     );
   }
 
@@ -228,20 +237,17 @@ class _OrderHistoryState extends State<OrderHistory> {
           // );
           ScaffoldMessenger.of(context).showSnackBar(
             SnackBar(
-              content: Text(
-                  '${blockedOrders!.length} orders unlocked'
-              ),
+              content: Text('${blockedOrders!.length} orders unlocked'),
             ),
           );
         } else {
           Navigator.push(
             context,
             MaterialPageRoute(
-              builder: (context) =>
-                  UpsellScreen(
-                    ordersLeft: ordersLeft,
-                    blockedOrders: blockedOrders!.length,
-                  ),
+              builder: (context) => UpsellScreen(
+                ordersLeft: ordersLeft,
+                blockedOrders: blockedOrders!.length,
+              ),
             ),
           );
         }
@@ -251,8 +257,8 @@ class _OrderHistoryState extends State<OrderHistory> {
         title: 'Locked orders',
         exception: PlatformException(
           code: 'ORDER_BUNDLED_PURCHASE_SUCCESS',
-          message:  'There are no locked orders at the moment.',
-          details:  'There are no locked orders at the moment',
+          message: 'There are no locked orders at the moment.',
+          details: 'There are no locked orders at the moment',
         ),
       ).show(context);
     }
@@ -268,18 +274,19 @@ class _OrderHistoryState extends State<OrderHistory> {
           widget.showLocked!
               ? 'Locked orders'
               : '${widget.restaurantName} Orders',
-          style: TextStyle(color: Theme.of(context).appBarTheme.backgroundColor),
+          style:
+              TextStyle(color: Theme.of(context).appBarTheme.backgroundColor),
         ),
         actions: [
           if (widget.showLocked!)
-          Padding(
-            padding: const EdgeInsets.only(right: 26.0),
-            child: IconButton(
-              iconSize: 24.0,
-              icon: Icon(Icons.lock_open),
-              onPressed: () => _unlockOrders(context),
+            Padding(
+              padding: const EdgeInsets.only(right: 26.0),
+              child: IconButton(
+                iconSize: 24.0,
+                icon: Icon(Icons.lock_open),
+                onPressed: () => _unlockOrders(context),
+              ),
             ),
-          ),
         ],
       ),
       body: _buildContents(context),
