@@ -6,13 +6,14 @@ import 'package:purchases_flutter/purchases_flutter.dart';
 enum SubscriptionType { Unsubscribed, Expired, Standard, Pro }
 
 class Subscription {
-  final PurchaserInfo? purchaserInfo;
+  final CustomerInfo? purchaserInfo;
   final Offerings? offerings;
 
   Subscription({this.purchaserInfo, this.offerings});
 
   SubscriptionType get subscriptionType {
-    if (purchaserInfo != null && purchaserInfo!.entitlements.active.isNotEmpty) {
+    if (purchaserInfo != null &&
+        purchaserInfo!.entitlements.active.isNotEmpty) {
       if (purchaserInfo!.entitlements.active.containsKey("pro")) {
         return SubscriptionType.Pro;
       } else {
@@ -27,22 +28,26 @@ class Subscription {
   String get subscriptionTypeString {
     String subCode;
     switch (subscriptionType) {
-      case SubscriptionType.Pro: {
-        subCode = 'Pro';
-      }
-      break;
-      case SubscriptionType.Standard: {
-        subCode = 'Standard';
-      }
-      break;
-      default: {
-        subCode = 'Unsubscribed';
-      }
+      case SubscriptionType.Pro:
+        {
+          subCode = 'Pro';
+        }
+        break;
+      case SubscriptionType.Standard:
+        {
+          subCode = 'Standard';
+        }
+        break;
+      default:
+        {
+          subCode = 'Unsubscribed';
+        }
     }
     return subCode;
   }
 
-  int get numberOfActiveSubscriptions => purchaserInfo!.activeSubscriptions.length;
+  int get numberOfActiveSubscriptions =>
+      purchaserInfo!.activeSubscriptions.length;
 
   String get firstSeen => purchaserInfo!.firstSeen;
 
@@ -54,7 +59,6 @@ class Subscription {
     }
     return null;
   }
-
 }
 
 abstract class IAPManagerBase {
@@ -66,139 +70,158 @@ abstract class IAPManagerBase {
   static String parseErrorCode(PurchasesErrorCode errorCode) {
     String errCode;
     switch (errorCode) {
-      case PurchasesErrorCode.unknownError: {
-        errCode = 'Unknown error';
-      }
-      break;
-      case PurchasesErrorCode.purchaseCancelledError: {
-        errCode = 'Purchase cancelled';
-      }
-      break;
-      case PurchasesErrorCode.storeProblemError: {
-        errCode = 'Store problem';
-      }
-      break;
-      case PurchasesErrorCode.purchaseNotAllowedError: {
-        errCode = 'Purchase not allowed';
-      }
-      break;
-      case PurchasesErrorCode.purchaseInvalidError: {
-        errCode = 'Purchase invalid';
-      }
-      break;
-      case PurchasesErrorCode.productNotAvailableForPurchaseError: {
-        errCode = 'Product not available for purchase';
-      }
-      break;
-      case PurchasesErrorCode.productAlreadyPurchasedError: {
-        errCode = 'Product already purchased';
-      }
-      break;
-      case PurchasesErrorCode.receiptAlreadyInUseError: {
-        errCode = 'Receipt already in use';
-      }
-      break;
-      case PurchasesErrorCode.invalidReceiptError: {
-        errCode = 'Invalid receipt';
-      }
-      break;
-      case PurchasesErrorCode.missingReceiptFileError: {
-        errCode = 'Missing receipt';
-      }
-      break;
-      case PurchasesErrorCode.networkError: {
-        errCode = 'Network error';
-      }
-      break;
-      case PurchasesErrorCode.invalidCredentialsError: {
-        errCode = 'Invalid credentials';
-      }
-      break;
-      case PurchasesErrorCode.unexpectedBackendResponseError: {
-        errCode = 'Unexpected backend response';
-      }
-      break;
-      case PurchasesErrorCode.receiptInUseByOtherSubscriberError: {
-        errCode = 'Receipt in use by other subscriber';
-      }
-      break;
-      case PurchasesErrorCode.invalidAppUserIdError: {
-        errCode = 'Invalid app user ID';
-      }
-      break;
-      case PurchasesErrorCode.operationAlreadyInProgressError: {
-        errCode = 'Operation already in progress';
-      }
-      break;
-      case PurchasesErrorCode.unknownBackendError: {
-        errCode = 'Unknown backend error';
-      }
-      break;
-      case PurchasesErrorCode.insufficientPermissionsError: {
-        errCode = 'Insufficient permissions';
-      }
-      break;
-      case PurchasesErrorCode.paymentPendingError: {
-        errCode = 'Payment pending';
-      }
-      break;
-      default: {
-        errCode = 'Other unknown error';
-      }
+      case PurchasesErrorCode.unknownError:
+        {
+          errCode = 'Unknown error';
+        }
+        break;
+      case PurchasesErrorCode.purchaseCancelledError:
+        {
+          errCode = 'Purchase cancelled';
+        }
+        break;
+      case PurchasesErrorCode.storeProblemError:
+        {
+          errCode = 'Store problem';
+        }
+        break;
+      case PurchasesErrorCode.purchaseNotAllowedError:
+        {
+          errCode = 'Purchase not allowed';
+        }
+        break;
+      case PurchasesErrorCode.purchaseInvalidError:
+        {
+          errCode = 'Purchase invalid';
+        }
+        break;
+      case PurchasesErrorCode.productNotAvailableForPurchaseError:
+        {
+          errCode = 'Product not available for purchase';
+        }
+        break;
+      case PurchasesErrorCode.productAlreadyPurchasedError:
+        {
+          errCode = 'Product already purchased';
+        }
+        break;
+      case PurchasesErrorCode.receiptAlreadyInUseError:
+        {
+          errCode = 'Receipt already in use';
+        }
+        break;
+      case PurchasesErrorCode.invalidReceiptError:
+        {
+          errCode = 'Invalid receipt';
+        }
+        break;
+      case PurchasesErrorCode.missingReceiptFileError:
+        {
+          errCode = 'Missing receipt';
+        }
+        break;
+      case PurchasesErrorCode.networkError:
+        {
+          errCode = 'Network error';
+        }
+        break;
+      case PurchasesErrorCode.invalidCredentialsError:
+        {
+          errCode = 'Invalid credentials';
+        }
+        break;
+      case PurchasesErrorCode.unexpectedBackendResponseError:
+        {
+          errCode = 'Unexpected backend response';
+        }
+        break;
+      case PurchasesErrorCode.receiptInUseByOtherSubscriberError:
+        {
+          errCode = 'Receipt in use by other subscriber';
+        }
+        break;
+      case PurchasesErrorCode.invalidAppUserIdError:
+        {
+          errCode = 'Invalid app user ID';
+        }
+        break;
+      case PurchasesErrorCode.operationAlreadyInProgressError:
+        {
+          errCode = 'Operation already in progress';
+        }
+        break;
+      case PurchasesErrorCode.unknownBackendError:
+        {
+          errCode = 'Unknown backend error';
+        }
+        break;
+      case PurchasesErrorCode.insufficientPermissionsError:
+        {
+          errCode = 'Insufficient permissions';
+        }
+        break;
+      case PurchasesErrorCode.paymentPendingError:
+        {
+          errCode = 'Payment pending';
+        }
+        break;
+      default:
+        {
+          errCode = 'Other unknown error';
+        }
     }
     return errCode;
   }
-
 }
 
 class IAPManagerMock implements IAPManagerBase {
   final String userID;
-  PurchaserInfo? _purchaserInfo;
+  CustomerInfo? _purchaserInfo;
   Offerings? _offerings;
   Subscription? _subscription;
-  StreamController<Subscription?> controller = StreamController<Subscription?>.broadcast();
+  StreamController<Subscription?> controller =
+      StreamController<Subscription?>.broadcast();
 
   IAPManagerMock({required this.userID}) {
     init();
   }
 
   Future<void> init() async {
-    _purchaserInfo = PurchaserInfo.fromJson(purchaserInfoTestDataU);
+    _purchaserInfo = CustomerInfo.fromJson(purchaserInfoTestDataU);
     _offerings = Offerings.fromJson(offeringsTestData);
     await Future.delayed(Duration(seconds: 3)); // Simulate slow network
     streamSubscription(pi: _purchaserInfo, of: _offerings);
   }
-  
+
   @override
   Stream<Subscription?> get onSubscriptionChanged => controller.stream;
 
   @override
   void purchasePackage(Package package) {
     print('Purchasing package: ${package.identifier}');
-    _purchaserInfo = PurchaserInfo.fromJson(purchaserInfoTestDataS);
+    _purchaserInfo = CustomerInfo.fromJson(purchaserInfoTestDataS);
     streamSubscription(pi: _purchaserInfo, of: _offerings);
   }
 
   @override
-  Future<void> purchaseProduct(String productIdentifier) async {
-  }
+  Future<void> purchaseProduct(String productIdentifier) async {}
 
   @override
   SubscriptionType get subscriptionType => _subscription!.subscriptionType;
 
-  void streamSubscription({PurchaserInfo? pi, Offerings? of}) {
+  void streamSubscription({CustomerInfo? pi, Offerings? of}) {
     _subscription = Subscription(purchaserInfo: pi, offerings: of);
     controller.add(_subscription);
   }
-
 }
 
 class IAPManager implements IAPManagerBase {
   final String? userID;
-  PurchaserInfo? _purchaserInfo;
+  CustomerInfo? _purchaserInfo;
   Offerings? _offerings;
   Subscription? _subscription;
-  StreamController<Subscription?> controller = StreamController<Subscription?>();
+  StreamController<Subscription?> controller =
+      StreamController<Subscription?>();
   static const String API_KEY = 'AeegEYeSxBwqtfZXZtbeMWVTOnAhyxiA';
 
   @override
@@ -210,22 +233,24 @@ class IAPManager implements IAPManagerBase {
 
   Future<void> init() async {
     try {
-      Purchases.setDebugLogsEnabled(false);
-      String? rcUserID = (userID == null || userID == '') ? 'anonymous' : userID;
-      await Purchases.setup(API_KEY, appUserId: rcUserID);
+      Purchases.setLogLevel(LogLevel.debug);
+      // String? rcUserID =
+      //     (userID == null || userID == '') ? 'anonymous' : userID;
+      await Purchases.configure(
+          PurchasesConfiguration(API_KEY)); //(API_KEY, appUserId: rcUserID);
       //await Purchases.setAllowSharingStoreAccount(false);
-      _purchaserInfo = await Purchases.getPurchaserInfo();
+      _purchaserInfo = await Purchases.getCustomerInfo();
       _offerings = await Purchases.getOfferings();
       streamSubscription(pi: _purchaserInfo, of: _offerings);
-      Purchases.addPurchaserInfoUpdateListener((pi) {
-            streamSubscription(pi: pi, of: _offerings);
-          });
+      Purchases.addCustomerInfoUpdateListener((pi) {
+        streamSubscription(pi: pi, of: _offerings);
+      });
     } catch (e) {
       print(e);
     }
   }
 
-  void streamSubscription({PurchaserInfo? pi, Offerings? of}) {
+  void streamSubscription({CustomerInfo? pi, Offerings? of}) {
     _subscription = Subscription(purchaserInfo: pi, offerings: of);
     controller.add(_subscription);
   }
@@ -248,7 +273,8 @@ class IAPManager implements IAPManagerBase {
   @override
   Future<void> purchaseProduct(String productIdentifier) async {
     try {
-      await Purchases.purchaseProduct(productIdentifier, type: PurchaseType.inapp);
+      await Purchases.purchaseProduct(productIdentifier,
+          type: PurchaseType.inapp);
       // Don't have to add anything to the stream as the listener above
       // will pick-up the subscription change and add it to the stream
     } on PlatformException catch (e) {
@@ -257,5 +283,4 @@ class IAPManager implements IAPManagerBase {
       rethrow;
     }
   }
-
 }
