@@ -1,466 +1,408 @@
-  import 'package:flutter/material.dart';
+import 'package:flutter/material.dart';
 
 class OrangeTheme {
-    static final ThemeData theme = ThemeData(
-      primarySwatch: Colors.orange,
-      brightness: Brightness.light,
-      primaryColor: Color(0xffff9800),
-      primaryColorLight: Color(0xffffe0b2),
-      primaryColorDark: Color(0xfff57c00),
-      //accentColor: Color(0xffff9800),
-      //accentColorBrightness: Brightness.light,
-      canvasColor: Color(0xfffafafa),
-      scaffoldBackgroundColor: Color(0xfffafafa),
-      bottomAppBarColor: Color(0xffffffff),
-      cardColor: Color(0xffffffff),
-      dividerColor: Color(0x1f000000),
-      highlightColor: Color(0x66bcbcbc),
-      splashColor: Color(0x66c8c8c8),
-      selectedRowColor: Color(0xfff5f5f5),
-      unselectedWidgetColor: Color(0x8a000000),
+  static final ThemeData theme = ThemeData(
+    primarySwatch: Colors.orange,
+    brightness: Brightness.light,
+    primaryColor: Color(0xffff9800),
+    primaryColorLight: Color(0xffffe0b2),
+    primaryColorDark: Color(0xfff57c00),
+    //accentColor: Color(0xffff9800),
+    //accentColorBrightness: Brightness.light,
+    canvasColor: Color(0xfffafafa),
+    scaffoldBackgroundColor: Color(0xfffafafa),
+    bottomAppBarTheme: BottomAppBarTheme(color: Color(0xffffffff)),
+    cardColor: Color(0xffffffff),
+    dividerColor: Color(0x1f000000),
+    highlightColor: Color(0x66bcbcbc),
+    splashColor: Color(0x66c8c8c8),
+    unselectedWidgetColor: Color(0x8a000000),
+    disabledColor: Color(0x61000000),
+    //buttonColor: Color(0xffe0e0e0),
+    secondaryHeaderColor: Color(0xfffff3e0),
+    //textSelectionColor: Color(0xffffcc80),
+    //cursorColor: Color(0xff4285f4),
+    //textSelectionHandleColor: Color(0xffffb74d),
+    backgroundColor: Color(0xffffcc80),
+    dialogBackgroundColor: Color(0xffffffff),
+    indicatorColor: Color(0xffff9800),
+    hintColor: Color(0x8a000000),
+    errorColor: Color(0xffd32f2f),
+    buttonTheme: ButtonThemeData(
+      textTheme: ButtonTextTheme.normal,
+      minWidth: 88.0,
+      height: 36.0,
+      padding: EdgeInsets.only(top: 0.0, bottom: 0.0, left: 16.0, right: 16.0),
+      shape: RoundedRectangleBorder(
+        side: BorderSide(
+          color: Color(0xff000000),
+          width: 0.0,
+          style: BorderStyle.none,
+        ),
+        borderRadius: BorderRadius.all(Radius.circular(2.0)),
+      ),
+      alignedDropdown: false,
+      buttonColor: Color(0xffe0e0e0),
       disabledColor: Color(0x61000000),
-      //buttonColor: Color(0xffe0e0e0),
-      toggleableActiveColor: Color(0xfffb8c00),
-      secondaryHeaderColor: Color(0xfffff3e0),
-      //textSelectionColor: Color(0xffffcc80),
-      //cursorColor: Color(0xff4285f4),
-      //textSelectionHandleColor: Color(0xffffb74d),
-      backgroundColor: Color(0xffffcc80),
-      dialogBackgroundColor: Color(0xffffffff),
-      indicatorColor: Color(0xffff9800),
-      hintColor: Color(0x8a000000),
-      errorColor: Color(0xffd32f2f),
-      buttonTheme: ButtonThemeData(
-        textTheme: ButtonTextTheme.normal,
-        minWidth: 88.0,
-        height: 36.0,
-        padding: EdgeInsets.only(
-            top: 0.0, bottom: 0.0, left: 16.0, right: 16.0),
-        shape: RoundedRectangleBorder(
-          side: BorderSide(
-            color: Color(0xff000000), width: 0.0, style: BorderStyle.none,),
-          borderRadius: BorderRadius.all(Radius.circular(2.0)),
-        )
-        ,
-        alignedDropdown: false,
-        buttonColor: Color(0xffe0e0e0),
-        disabledColor: Color(0x61000000),
-        highlightColor: Color(0x29000000),
-        splashColor: Color(0x1f000000),
-        colorScheme: ColorScheme(
-          primary: Color(0xffff9800),
-          secondary: Color(0xffff9800),
-          surface: Color(0xffffffff),
-          background: Color(0xffffcc80),
-          error: Color(0xffd32f2f),
-          onPrimary: Color(0xff000000),
-          onSecondary: Color(0xff000000),
-          onSurface: Color(0xff000000),
-          onBackground: Color(0xff000000),
-          onError: Color(0xffffffff),
-          brightness: Brightness.light,
-        ),
-      ),
-      textTheme: TextTheme(
-        headline1: TextStyle(
-          color: Color(0x8a000000),
-          fontSize: 96.0,
-          fontWeight: FontWeight.w300,
-          fontStyle: FontStyle.normal,
-        ),
-        headline2: TextStyle(
-          color: Color(0x8a000000),
-          fontSize: 60.0,
-          fontWeight: FontWeight.w300,
-          fontStyle: FontStyle.normal,
-        ),
-        headline3: TextStyle(
-          color: Color(0x8a000000),
-          fontSize: 48.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        headline4: TextStyle(
-          color: Color(0x8a000000),
-          fontSize: 34.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        headline5: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 24.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        headline6: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 20.0,
-          fontWeight: FontWeight.w500,
-          fontStyle: FontStyle.normal,
-        ),
-        subtitle1: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 16.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        bodyText1: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 14.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        bodyText2: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 16.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        caption: TextStyle(
-          color: Color(0x8a000000),
-          fontSize: 12.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        button: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 14.0,
-          fontWeight: FontWeight.w500,
-          fontStyle: FontStyle.normal,
-        ),
-        subtitle2: TextStyle(
-          color: Color(0xff000000),
-          fontSize: 14.0,
-          fontWeight: FontWeight.w500,
-          fontStyle: FontStyle.normal,
-        ),
-        overline: TextStyle(
-          color: Color(0xff000000),
-          fontSize: 10.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-      ),
-      primaryTextTheme: TextTheme(
-        headline1: TextStyle(
-          color: Color(0x8a000000),
-          fontSize: 96.0,
-          fontWeight: FontWeight.w300,
-          fontStyle: FontStyle.normal,
-        ),
-        headline2: TextStyle(
-          color: Color(0x8a000000),
-          fontSize: 60.0,
-          fontWeight: FontWeight.w300,
-          fontStyle: FontStyle.normal,
-        ),
-        headline3: TextStyle(
-          color: Color(0x8a000000),
-          fontSize: 48.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        headline4: TextStyle(
-          color: Color(0x8a000000),
-          fontSize: 34.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        headline5: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 24.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        headline6: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 20.0,
-          fontWeight: FontWeight.w500,
-          fontStyle: FontStyle.normal,
-        ),
-        subtitle1: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 16.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        bodyText1: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 14.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        bodyText2: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 16.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        caption: TextStyle(
-          color: Color(0x8a000000),
-          fontSize: 12.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        button: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 14.0,
-          fontWeight: FontWeight.w500,
-          fontStyle: FontStyle.normal,
-        ),
-        subtitle2: TextStyle(
-          color: Color(0xff000000),
-          fontSize: 14.0,
-          fontWeight: FontWeight.w500,
-          fontStyle: FontStyle.normal,
-        ),
-        overline: TextStyle(
-          color: Color(0xff000000),
-          fontSize: 10.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-      ),
-      // accentTextTheme: TextTheme(
-      //   headline1: TextStyle(
-      //     color: Color(0x8a000000),
-      //     fontSize: 96.0,
-      //     fontWeight: FontWeight.w300,
-      //     fontStyle: FontStyle.normal,
-      //   ),
-      //   headline2: TextStyle(
-      //     color: Color(0x8a000000),
-      //     fontSize: 60.0,
-      //     fontWeight: FontWeight.w300,
-      //     fontStyle: FontStyle.normal,
-      //   ),
-      //   headline3: TextStyle(
-      //     color: Color(0x8a000000),
-      //     fontSize: 48.0,
-      //     fontWeight: FontWeight.w400,
-      //     fontStyle: FontStyle.normal,
-      //   ),
-      //   headline4: TextStyle(
-      //     color: Color(0x8a000000),
-      //     fontSize: 34.0,
-      //     fontWeight: FontWeight.w400,
-      //     fontStyle: FontStyle.normal,
-      //   ),
-      //   headline5: TextStyle(
-      //     color: Color(0xdd000000),
-      //     fontSize: 24.0,
-      //     fontWeight: FontWeight.w400,
-      //     fontStyle: FontStyle.normal,
-      //   ),
-      //   headline6: TextStyle(
-      //     color: Color(0xdd000000),
-      //     fontSize: 20.0,
-      //     fontWeight: FontWeight.w500,
-      //     fontStyle: FontStyle.normal,
-      //   ),
-      //   subtitle1: TextStyle(
-      //     color: Color(0xdd000000),
-      //     fontSize: 16.0,
-      //     fontWeight: FontWeight.w400,
-      //     fontStyle: FontStyle.normal,
-      //   ),
-      //   bodyText1: TextStyle(
-      //     color: Color(0xdd000000),
-      //     fontSize: 14.0,
-      //     fontWeight: FontWeight.w400,
-      //     fontStyle: FontStyle.normal,
-      //   ),
-      //   bodyText2: TextStyle(
-      //     color: Color(0xdd000000),
-      //     fontSize: 16.0,
-      //     fontWeight: FontWeight.w400,
-      //     fontStyle: FontStyle.normal,
-      //   ),
-      //   caption: TextStyle(
-      //     color: Color(0x8a000000),
-      //     fontSize: 12.0,
-      //     fontWeight: FontWeight.w400,
-      //     fontStyle: FontStyle.normal,
-      //   ),
-      //   button: TextStyle(
-      //     color: Color(0xdd000000),
-      //     fontSize: 14.0,
-      //     fontWeight: FontWeight.w500,
-      //     fontStyle: FontStyle.normal,
-      //   ),
-      //   subtitle2: TextStyle(
-      //     color: Color(0xff000000),
-      //     fontSize: 14.0,
-      //     fontWeight: FontWeight.w500,
-      //     fontStyle: FontStyle.normal,
-      //   ),
-      //   overline: TextStyle(
-      //     color: Color(0xff000000),
-      //     fontSize: 10.0,
-      //     fontWeight: FontWeight.w400,
-      //     fontStyle: FontStyle.normal,
-      //   ),
-      // ),
-      inputDecorationTheme: InputDecorationTheme(
-        labelStyle: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 16.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        helperStyle: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 16.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        hintStyle: TextStyle(
-          color: Color(0xffff9800),
-          fontSize: 16.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        errorStyle: TextStyle(
-          color: Color(0xffd32f2f),
-          fontSize: 16.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        errorMaxLines: null,
-        floatingLabelBehavior: FloatingLabelBehavior.never,
-        isDense: false,
-        contentPadding: EdgeInsets.only(
-            top: 12.0, bottom: 12.0, left: 0.0, right: 0.0),
-        isCollapsed: false,
-        prefixStyle: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 16.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        suffixStyle: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 16.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        counterStyle: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 16.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        filled: false,
-        fillColor: Color(0x00000000),
-        errorBorder: UnderlineInputBorder(
-          borderSide: BorderSide(
-            color: Color(0xff000000), width: 1.0, style: BorderStyle.solid,),
-          borderRadius: BorderRadius.all(Radius.circular(4.0)),
-        ),
-        focusedBorder: UnderlineInputBorder(
-          borderSide: BorderSide(
-            color: Color(0xff000000), width: 1.0, style: BorderStyle.solid,),
-          borderRadius: BorderRadius.all(Radius.circular(4.0)),
-        ),
-        focusedErrorBorder: UnderlineInputBorder(
-          borderSide: BorderSide(
-            color: Color(0xff000000), width: 1.0, style: BorderStyle.solid,),
-          borderRadius: BorderRadius.all(Radius.circular(4.0)),
-        ),
-        disabledBorder: UnderlineInputBorder(
-          borderSide: BorderSide(
-            color: Color(0xff000000), width: 1.0, style: BorderStyle.solid,),
-          borderRadius: BorderRadius.all(Radius.circular(4.0)),
-        ),
-        enabledBorder: UnderlineInputBorder(
-          borderSide: BorderSide(
-            color: Color(0xff000000), width: 1.0, style: BorderStyle.solid,),
-          borderRadius: BorderRadius.all(Radius.circular(4.0)),
-        ),
-        border: UnderlineInputBorder(
-          borderSide: BorderSide(
-            color: Color(0xff000000), width: 1.0, style: BorderStyle.solid,),
-          borderRadius: BorderRadius.all(Radius.circular(4.0)),
-        ),
-      ),
-      iconTheme: IconThemeData(
-        color: Color(0xdd000000),
-        opacity: 1.0,
-        size: 24.0,
-      ),
-      primaryIconTheme: IconThemeData(
-        color: Color(0xff000000),
-        opacity: 1.0,
-        size: 24.0,
-      ),
-      floatingActionButtonTheme: FloatingActionButtonThemeData(
-        foregroundColor: Color(0xff000000),
-      ),
-      // accentIconTheme: IconThemeData(
-      //   color: Color(0xff000000),
-      //   opacity: 1.0,
-      //   size: 24.0,
-      // ),
-      sliderTheme: SliderThemeData(
-        activeTrackColor: Color(0xffff9800),
-        inactiveTrackColor: Color(0x3dff9800),
-        disabledActiveTrackColor: Color(0x52f57c00),
-        disabledInactiveTrackColor: Color(0x1ff57c00),
-        activeTickMarkColor: Color(0x8affe0b2),
-        inactiveTickMarkColor: Color(0x8aff9800),
-        disabledActiveTickMarkColor: Color(0x1fffe0b2),
-        disabledInactiveTickMarkColor: Color(0x1ff57c00),
-        thumbColor: Color(0xffff9800),
-        disabledThumbColor: Color(0x52f57c00),
-        thumbShape: RoundSliderThumbShape(),
-        overlayColor: Color(0x29ff9800),
-        valueIndicatorColor: Color(0xffff9800),
-        valueIndicatorShape: PaddleSliderValueIndicatorShape(),
-        showValueIndicator: ShowValueIndicator.onlyForDiscrete,
-        valueIndicatorTextStyle: TextStyle(
-          color: Color(0xdd000000),
-          fontSize: 14.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-      ),
-      tabBarTheme: TabBarTheme(
-        indicatorSize: TabBarIndicatorSize.tab,
-        labelColor: Color(0xdd000000),
-        unselectedLabelColor: Color(0xb2000000),
-      ),
-      chipTheme: ChipThemeData(
-        backgroundColor: Color(0x1f000000),
+      highlightColor: Color(0x29000000),
+      splashColor: Color(0x1f000000),
+      colorScheme: ColorScheme(
+        primary: Color(0xffff9800),
+        secondary: Color(0xffff9800),
+        surface: Color(0xffffffff),
+        background: Color(0xffffcc80),
+        error: Color(0xffd32f2f),
+        onPrimary: Color(0xff000000),
+        onSecondary: Color(0xff000000),
+        onSurface: Color(0xff000000),
+        onBackground: Color(0xff000000),
+        onError: Color(0xffffffff),
         brightness: Brightness.light,
-        deleteIconColor: Color(0xde000000),
-        disabledColor: Color(0x0c000000),
-        labelPadding: EdgeInsets.only(
-            top: 0.0, bottom: 0.0, left: 8.0, right: 8.0),
-        labelStyle: TextStyle(
-          color: Color(0xde000000),
-          fontSize: 14.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        padding: EdgeInsets.only(top: 4.0, bottom: 4.0, left: 4.0, right: 4.0),
-        secondaryLabelStyle: TextStyle(
-          color: Color(0x3d000000),
-          fontSize: 14.0,
-          fontWeight: FontWeight.w400,
-          fontStyle: FontStyle.normal,
-        ),
-        secondarySelectedColor: Color(0x3dff9800),
-        selectedColor: Color(0x3d000000),
-        shape: StadiumBorder(side: BorderSide(
-          color: Color(0xff000000), width: 0.0, style: BorderStyle.none,)),
       ),
-      dialogTheme: DialogTheme(
-          shape: RoundedRectangleBorder(
-            side: BorderSide(
-              color: Color(0xff000000), width: 0.0, style: BorderStyle.none,),
-            borderRadius: BorderRadius.all(Radius.circular(0.0)),
-          )
-
+    ),
+    textTheme: TextTheme(
+      displayLarge: TextStyle(
+        color: Color(0x8a000000),
+        fontSize: 96.0,
+        fontWeight: FontWeight.w300,
+        fontStyle: FontStyle.normal,
       ),
-    );
+      displayMedium: TextStyle(
+        color: Color(0x8a000000),
+        fontSize: 60.0,
+        fontWeight: FontWeight.w300,
+        fontStyle: FontStyle.normal,
+      ),
+      displaySmall: TextStyle(
+        color: Color(0x8a000000),
+        fontSize: 48.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      headlineMedium: TextStyle(
+        color: Color(0x8a000000),
+        fontSize: 34.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      headlineSmall: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 24.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      titleLarge: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 20.0,
+        fontWeight: FontWeight.w500,
+        fontStyle: FontStyle.normal,
+      ),
+      titleMedium: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 16.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      bodyLarge: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 14.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      bodyMedium: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 16.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      bodySmall: TextStyle(
+        color: Color(0x8a000000),
+        fontSize: 12.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      labelLarge: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 14.0,
+        fontWeight: FontWeight.w500,
+        fontStyle: FontStyle.normal,
+      ),
+      titleSmall: TextStyle(
+        color: Color(0xff000000),
+        fontSize: 14.0,
+        fontWeight: FontWeight.w500,
+        fontStyle: FontStyle.normal,
+      ),
+      labelSmall: TextStyle(
+        color: Color(0xff000000),
+        fontSize: 10.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+    ),
+    primaryTextTheme: TextTheme(
+      displayLarge: TextStyle(
+        color: Color(0x8a000000),
+        fontSize: 96.0,
+        fontWeight: FontWeight.w300,
+        fontStyle: FontStyle.normal,
+      ),
+      displayMedium: TextStyle(
+        color: Color(0x8a000000),
+        fontSize: 60.0,
+        fontWeight: FontWeight.w300,
+        fontStyle: FontStyle.normal,
+      ),
+      displaySmall: TextStyle(
+        color: Color(0x8a000000),
+        fontSize: 48.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      headlineMedium: TextStyle(
+        color: Color(0x8a000000),
+        fontSize: 34.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      headlineSmall: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 24.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      titleLarge: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 20.0,
+        fontWeight: FontWeight.w500,
+        fontStyle: FontStyle.normal,
+      ),
+      titleMedium: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 16.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      bodyLarge: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 14.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      bodyMedium: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 16.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      bodySmall: TextStyle(
+        color: Color(0x8a000000),
+        fontSize: 12.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      labelLarge: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 14.0,
+        fontWeight: FontWeight.w500,
+        fontStyle: FontStyle.normal,
+      ),
+      titleSmall: TextStyle(
+        color: Color(0xff000000),
+        fontSize: 14.0,
+        fontWeight: FontWeight.w500,
+        fontStyle: FontStyle.normal,
+      ),
+      labelSmall: TextStyle(
+        color: Color(0xff000000),
+        fontSize: 10.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+    ),
+    inputDecorationTheme: InputDecorationTheme(
+      labelStyle: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 16.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      helperStyle: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 16.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      hintStyle: TextStyle(
+        color: Color(0xffff9800),
+        fontSize: 16.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      errorStyle: TextStyle(
+        color: Color(0xffd32f2f),
+        fontSize: 16.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      errorMaxLines: null,
+      floatingLabelBehavior: FloatingLabelBehavior.never,
+      isDense: false,
+      contentPadding:
+          EdgeInsets.only(top: 12.0, bottom: 12.0, left: 0.0, right: 0.0),
+      isCollapsed: false,
+      prefixStyle: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 16.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      suffixStyle: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 16.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      counterStyle: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 16.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      filled: false,
+      fillColor: Color(0x00000000),
+      errorBorder: UnderlineInputBorder(
+        borderSide: BorderSide(
+          color: Color(0xff000000),
+          width: 1.0,
+          style: BorderStyle.solid,
+        ),
+        borderRadius: BorderRadius.all(Radius.circular(4.0)),
+      ),
+      focusedBorder: UnderlineInputBorder(
+        borderSide: BorderSide(
+          color: Color(0xff000000),
+          width: 1.0,
+          style: BorderStyle.solid,
+        ),
+        borderRadius: BorderRadius.all(Radius.circular(4.0)),
+      ),
+      focusedErrorBorder: UnderlineInputBorder(
+        borderSide: BorderSide(
+          color: Color(0xff000000),
+          width: 1.0,
+          style: BorderStyle.solid,
+        ),
+        borderRadius: BorderRadius.all(Radius.circular(4.0)),
+      ),
+      disabledBorder: UnderlineInputBorder(
+        borderSide: BorderSide(
+          color: Color(0xff000000),
+          width: 1.0,
+          style: BorderStyle.solid,
+        ),
+        borderRadius: BorderRadius.all(Radius.circular(4.0)),
+      ),
+      enabledBorder: UnderlineInputBorder(
+        borderSide: BorderSide(
+          color: Color(0xff000000),
+          width: 1.0,
+          style: BorderStyle.solid,
+        ),
+        borderRadius: BorderRadius.all(Radius.circular(4.0)),
+      ),
+      border: UnderlineInputBorder(
+        borderSide: BorderSide(
+          color: Color(0xff000000),
+          width: 1.0,
+          style: BorderStyle.solid,
+        ),
+        borderRadius: BorderRadius.all(Radius.circular(4.0)),
+      ),
+    ),
+    iconTheme: IconThemeData(
+      color: Color(0xdd000000),
+      opacity: 1.0,
+      size: 24.0,
+    ),
+    primaryIconTheme: IconThemeData(
+      color: Color(0xff000000),
+      opacity: 1.0,
+      size: 24.0,
+    ),
+    floatingActionButtonTheme: FloatingActionButtonThemeData(
+      foregroundColor: Color(0xff000000),
+    ),
+    // accentIconTheme: IconThemeData(
+    //   color: Color(0xff000000),
+    //   opacity: 1.0,
+    //   size: 24.0,
+    // ),
+    sliderTheme: SliderThemeData(
+      activeTrackColor: Color(0xffff9800),
+      inactiveTrackColor: Color(0x3dff9800),
+      disabledActiveTrackColor: Color(0x52f57c00),
+      disabledInactiveTrackColor: Color(0x1ff57c00),
+      activeTickMarkColor: Color(0x8affe0b2),
+      inactiveTickMarkColor: Color(0x8aff9800),
+      disabledActiveTickMarkColor: Color(0x1fffe0b2),
+      disabledInactiveTickMarkColor: Color(0x1ff57c00),
+      thumbColor: Color(0xffff9800),
+      disabledThumbColor: Color(0x52f57c00),
+      thumbShape: RoundSliderThumbShape(),
+      overlayColor: Color(0x29ff9800),
+      valueIndicatorColor: Color(0xffff9800),
+      valueIndicatorShape: PaddleSliderValueIndicatorShape(),
+      showValueIndicator: ShowValueIndicator.onlyForDiscrete,
+      valueIndicatorTextStyle: TextStyle(
+        color: Color(0xdd000000),
+        fontSize: 14.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+    ),
+    tabBarTheme: TabBarTheme(
+      indicatorSize: TabBarIndicatorSize.tab,
+      labelColor: Color(0xdd000000),
+      unselectedLabelColor: Color(0xb2000000),
+    ),
+    chipTheme: ChipThemeData(
+      backgroundColor: Color(0x1f000000),
+      brightness: Brightness.light,
+      deleteIconColor: Color(0xde000000),
+      disabledColor: Color(0x0c000000),
+      labelPadding:
+          EdgeInsets.only(top: 0.0, bottom: 0.0, left: 8.0, right: 8.0),
+      labelStyle: TextStyle(
+        color: Color(0xde000000),
+        fontSize: 14.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      padding: EdgeInsets.only(top: 4.0, bottom: 4.0, left: 4.0, right: 4.0),
+      secondaryLabelStyle: TextStyle(
+        color: Color(0x3d000000),
+        fontSize: 14.0,
+        fontWeight: FontWeight.w400,
+        fontStyle: FontStyle.normal,
+      ),
+      secondarySelectedColor: Color(0x3dff9800),
+      selectedColor: Color(0x3d000000),
+      shape: StadiumBorder(
+          side: BorderSide(
+        color: Color(0xff000000),
+        width: 0.0,
+        style: BorderStyle.none,
+      )),
+    ),
+    dialogTheme: DialogTheme(
+        shape: RoundedRectangleBorder(
+      side: BorderSide(
+        color: Color(0xff000000),
+        width: 0.0,
+        style: BorderStyle.none,
+      ),
+      borderRadius: BorderRadius.all(Radius.circular(0.0)),
+    )),
+  );
 }
